test(vetting): Wave 6 — split test_2d_remaining_features.h by family
The cross-cutting 2D leftover file (7 functions across 3 families that
share one compute pass) split by registry target_test:

- test_remaining2d_common.h (git mv from the old file): #pragma-once
  shared fixture + assert helpers + 4 golden maps.
- test_morphology_regression.h: +5 functions (erosion-complement,
  caliper feret/martin/nassenstein, chord stats, chord angles,
  polygonality/hexagonality).
- test_intensity_histogram_regression.h: +radial distribution
  (FRAC_AT_D/MEAN_FRAC/RADIAL_CV).
- test_zernike_regression.h: new 1-function family file for ZERNIKE2D.

Function names kept verbatim so every TEST() in test_all.cc still
registers. 46 registry rows repointed off the old filename.

VERIFIED: full gtest suite 696/696 — no tests dropped; registry
validation rc=0.
