+++ tests/test_all.cc
@@ -16,7 +16,8 @@
 #include "test_morphology_fraclac.h"
 #include "test_moments_skimage.h"
 #include "test_moments_regression.h"
-#include "test_2d_remaining_features.h"
+#include "test_remaining2d_common.h"
+#include "test_zernike_regression.h"
 #include "test_neighbor_regression.h"
 #include "test_initialization.h"
 #include "test_glcm_ibsi.h"
