+++ tests/test_intensity_histogram_regression.h
@@ -6,6 +6,7 @@
 #include "../src/nyx/dataset.h"
 #include "../src/nyx/roi_cache.h"
 #include "../src/nyx/features/intensity_histogram.h"
+#include "test_remaining2d_common.h"
 #include "../src/nyx/features/pixel.h"
 #include "test_data.h"
 #include "test_main_nyxus.h"
@@ -276,4 +277,21 @@ void test_ih_required_predicate()
     ASSERT_FALSE(IntensityHistogramFeatures::required(fs));
     fs.enableFeature((int)Feature2D::IH_ENTROPY_VAL);
     ASSERT_TRUE(IntensityHistogramFeatures::required(fs));
+}
+
+// ---------------------------------------------------------------------------------------------------
+// Migrated from test_2d_remaining_features.h (Wave 6): radial intensity distribution (FRAC_AT_D,
+// MEAN_FRAC, RADIAL_CV). Registry target_test = test_intensity_histogram_regression.h; oracle is
+// cellprofiler MeasureObjectIntensityDistribution (RadialDistribution_*), still regression-snapshot
+// pending wiring. Shared fixture/oracle-data lives in test_remaining2d_common.h.
+// ---------------------------------------------------------------------------------------------------
+
+void test_remaining2d_unvetted_no_direct_oracle_radial_distribution_features()
+{
+	std::vector<std::vector<double>> fvals;
+	calculate_remaining2d_shape_feature_values(fvals);
+
+	assert_unvetted_no_direct_oracle_remaining2d_vector_feature(fvals, Nyxus::Feature2D::FRAC_AT_D, "FRAC_AT_D");
+	assert_unvetted_no_direct_oracle_remaining2d_vector_feature(fvals, Nyxus::Feature2D::MEAN_FRAC, "MEAN_FRAC");
+	assert_unvetted_no_direct_oracle_remaining2d_vector_feature(fvals, Nyxus::Feature2D::RADIAL_CV, "RADIAL_CV");
 }
